requirements.txt
Flask==3.0.3
gunicorn==21.2.0
gspread==6.1.2
google-auth==2.33.0
